async handle action, reusable swap
