+++ src/lib/actions/AddLiquidityToOciswap.svelte
@@ -83,7 +83,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
