+++ src/lib/actions/AddYourLSUsToCaviarnine.svelte
@@ -91,7 +91,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
