+++ src/lib/actions/AirdropFungibleCoins.svelte
@@ -60,7 +60,7 @@
     }
   }
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
