+++ src/lib/actions/DefiplazaSwapCoins.svelte
@@ -1,20 +1,17 @@
 <script lang="ts">
-  import { afterUpdate, onDestroy, onMount } from "svelte";
-  import { defiplaza_component, find_fungible_symbol } from "../../content";
+  import { defiplaza_component } from "../../content";
   import { defiplaza_listed_coins } from "../../defiplaza";
   import PrecisionNumber from "../PrecisionNumber";
   import commands from "../commands";
-  import AddActionButton from "../shared/AddActionButton.svelte";
-  import QuantityInput from "../shared/QuantityInput.svelte";
+  import SwapFungibles from "../shared/SwapFungibles.svelte";
   import type { WalletFungible } from "../stores/accounts";
   import { UNKNOWN_QUANTITY } from "../stores/accounts";
   import {
     CANNOT_PROCEED_WITH_UNKNOWN_QUANTITY,
     NO_COINS_TO_SEND,
     NO_DEFIPLAZA_COINS_ON_WORTOP,
+    NO_QUANTITY,
     actionError,
-    validateAvailableFungibles,
-    validateQuantity,
     validationErrors,
   } from "../stores/errors";
   import { bucketNumber, manifest } from "../stores/transaction";
@@ -25,69 +22,13 @@
   let sendFungibleAddress = "";
   let quantity = "";
   let allQuantity = true;
-  let maxQuantity: PrecisionNumber | undefined = undefined;
-  // fungible address -> fungible symbol
-  let receiveFungibles: Map<string, string> = new Map();
   let receiveFungibleAddress = "";
 
-  onMount(() => {
-    actionError.set("");
-  });
+  let listedFungibleAddresses = Object.keys(defiplaza_listed_coins);
 
-  function updateSelectors() {
-    availableFungibles = worktop.filterFungibles(
-      Object.keys(defiplaza_listed_coins)
-    );
-    if (!availableFungibles.has(sendFungibleAddress)) {
-      sendFungibleAddress = "";
-    }
-    if (sendFungibleAddress === "" && availableFungibles.size > 0) {
-      sendFungibleAddress = availableFungibles.keys().next().value;
-    }
-
-    receiveFungibles.clear();
-    for (let resourceAddress of Object.keys(defiplaza_listed_coins)) {
-      if (resourceAddress !== sendFungibleAddress) {
-        receiveFungibles.set(
-          resourceAddress,
-          find_fungible_symbol(resourceAddress)
-        );
-      }
-    }
-
-    // update svelte state
-    receiveFungibles = receiveFungibles;
-    if (
-      receiveFungibleAddress === "" ||
-      !receiveFungibles.has(receiveFungibleAddress)
-    ) {
-      receiveFungibleAddress = receiveFungibles.keys().next().value;
-    }
-    const prevMaxQuantity = maxQuantity;
-    const amount = availableFungibles.get(sendFungibleAddress)?.amount;
-    if (amount !== UNKNOWN_QUANTITY) {
-      maxQuantity = amount;
-    }
-    if (
-      maxQuantity !== undefined &&
-      (quantity === "" || prevMaxQuantity !== maxQuantity)
-    ) {
-      quantity = maxQuantity.toString();
-    }
+  $: if ($worktop) {
+    availableFungibles = worktop.filterFungibles(listedFungibleAddresses);
   }
-
-  afterUpdate(() => {
-    updateSelectors();
-    validateQuantity(quantity, maxQuantity);
-    validateAvailableFungibles(
-      availableFungibles,
-      NO_DEFIPLAZA_COINS_ON_WORTOP
-    );
-  });
-
-  onDestroy(() => {
-    validationErrors.clear();
-  });
 
   async function handleAddAction() {
     actionError.set("");
@@ -95,8 +36,11 @@
       return;
     }
 
-    if (quantity === "" || sendFungibleAddress === "") {
+    if (sendFungibleAddress === "") {
       throw new Error(NO_COINS_TO_SEND);
+    }
+    if (!allQuantity && quantity === "") {
+      throw new Error(NO_QUANTITY);
     }
     const sendFungible = availableFungibles.get(sendFungibleAddress);
     if (sendFungible === undefined) {
@@ -156,44 +100,13 @@
   }
 </script>
 
-<div class="flex space-x-12 w-full place-items-end">
-  <div class="form-control flex-grow space-y-2">
-    <label class="label">
-      <span class="label-text">Coin to send</span>
-      <select
-        class="select select-secondary select-sm w-3/5 text-end"
-        bind:value={sendFungibleAddress}
-      >
-        {#each Array.from(availableFungibles.values()) as sendFungible}
-          <option value={sendFungible.address}>
-            {sendFungible.symbol}
-          </option>
-        {/each}
-      </select>
-    </label>
-    <div class="label space-x-2 !mt-0 pt-0">
-      <label class="label cursor-pointer space-x-4 px-0">
-        <span class="label-text">all</span>
-        <input class="checkbox" type="checkbox" bind:checked={allQuantity} />
-      </label>
-      <QuantityInput bind:value={quantity} hidden={allQuantity} />
-    </div>
-
-    <label class="label">
-      <span class="label-text">Coin to receive</span>
-      <select
-        class="select select-secondary select-sm w-3/5 text-end"
-        bind:value={receiveFungibleAddress}
-      >
-        {#each Array.from(receiveFungibles) as [address, symbol]}
-          <option value={address}>
-            {symbol}
-          </option>
-        {/each}
-      </select>
-    </label>
-  </div>
-  <div>
-    <AddActionButton {handleAddAction} />
-  </div>
-</div>
+<SwapFungibles
+  noFungiblesOnWorktopError={NO_DEFIPLAZA_COINS_ON_WORTOP}
+  {availableFungibles}
+  bind:sendFungibleAddress
+  bind:quantity
+  bind:allQuantity
+  bind:receiveFungibleAddress
+  {listedFungibleAddresses}
+  {handleAddAction}
+/>
