+++ src/lib/actions/GableGetFlashloan.svelte
@@ -31,7 +31,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
