+++ src/lib/actions/GableProvideLiquidity.svelte
@@ -37,7 +37,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
