+++ src/lib/actions/GableRepayFlashloan.svelte
@@ -38,7 +38,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
