+++ src/lib/actions/RetrieveLSUsFromCaviarnine.svelte
@@ -90,7 +90,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
