+++ src/lib/actions/SendCoinsToAccount.svelte
@@ -51,7 +51,7 @@
     }
   }
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
