+++ src/lib/actions/SendCoinsToSelf.svelte
@@ -47,7 +47,7 @@
     }
   }
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
