+++ src/lib/actions/StakeYourXRDs.svelte
@@ -46,7 +46,7 @@
     quantityXRD = avaiableXRDs.amount.toString();
   }
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
