+++ src/lib/actions/SwapLSUsOnCaviarnine.svelte
@@ -109,7 +109,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
