+++ src/lib/actions/TakeCoinsFromSelf.svelte
@@ -42,7 +42,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
