+++ src/lib/actions/UnstakeYourLSUs.svelte
@@ -50,7 +50,7 @@
     quantity = maxQuantity.toString();
   }
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
