+++ src/lib/actions/WithdrawLiquidityFromOciswap.svelte
@@ -71,7 +71,7 @@
     validationErrors.clear();
   });
 
-  function handleAddAction() {
+  async function handleAddAction() {
     actionError.set("");
     if ($validationErrors.size > 0) {
       return;
